[@svelteui/core]: small fixes for timeline and grid
