+++ packages/svelteui-core/src/lib/components/Grid/Col/Col.styles.ts
@@ -1,8 +1,5 @@
 import { createStyles } from '$lib/styles';
-import type { Writable } from 'svelte/store';
 import type { DefaultProps, SvelteUINumberSize, SvelteUITheme } from '$lib/styles';
-
-export type GridContext = Writable<{ cols: number; grow: boolean; spacing: SvelteUINumberSize }>;
 
 export interface ColProps extends DefaultProps {
 	span?: number;
