+++ packages/svelteui-core/src/lib/components/Grid/Grid.styles.ts
@@ -1,5 +1,8 @@
 import { createStyles } from '$lib/styles';
+import type { Writable } from 'svelte/store';
 import type { CSS, DefaultProps, SvelteUINumberSize } from '$lib/styles';
+
+export type GridContext = Writable<{ cols: number; grow: boolean; spacing: SvelteUINumberSize }>;
 
 export interface GridProps extends DefaultProps {
 	align?: CSS['alignItems'];
